+++ archai/supergraph/algos/darts/bilevel_optimizer.py
@@ -12,8 +12,8 @@ from torch.optim.optimizer import Optimizer
 from archai.common.config import Config
 from archai.common import utils
 from archai.supergraph.nas.model import Model
-from archai.common.logger import Logger
-logger = Logger(source=__name__)
+
+from archai.common.common import logger
 from archai.common.utils import zip_eq
 from archai.supergraph.utils import ml_utils
