+++ archai/supergraph/algos/divnas/divnas_finalizers.py
@@ -10,8 +10,8 @@ from torch import nn
 import numpy as np
 
 from archai.common.common import get_conf
-from archai.common.logger import Logger
-logger = Logger(source=__name__)
+
+from archai.common.common import logger
 from archai.supergraph.datasets.data import get_data
 from archai.supergraph.nas.model import Model
 from archai.supergraph.nas.cell import Cell
