+++ archai/supergraph/algos/petridish/petridish_utils.py
@@ -18,8 +18,8 @@ import matplotlib.pyplot as plt
 from archai.supergraph.nas.model_desc import ConvMacroParams, CellDesc, CellType, OpDesc, \
                                   EdgeDesc, TensorShape, TensorShapes, NodeDesc, ModelDesc
 from archai.supergraph.utils.metrics import Metrics
-from archai.common.logger import Logger
-logger = Logger(source=__name__)
+
+from archai.common.common import logger
 from archai.common import utils
 
 class JobStage(Enum):
