+++ archai/supergraph/algos/xnas/xnas_arch_trainer.py
@@ -21,8 +21,8 @@ from archai.supergraph.nas.model import Model
 from archai.supergraph.nas.model_desc import CellType
 from archai.supergraph.utils import ml_utils
 from archai.supergraph.utils.checkpoint import CheckPoint
-from archai.common.logger import Logger
-logger = Logger(source=__name__)
+
+from archai.common.common import logger
 from archai.supergraph.datasets import data
 from archai.common.common import get_conf
 from archai.supergraph.algos.xnas.xnas_op import XnasOp
