+++ archai/supergraph/datasets/providers/imagenet_folder.py
@@ -11,9 +11,9 @@ from torchvision.transforms import transforms
 from torch.utils.data.dataset import Dataset
 from torchvision.datasets.utils import check_integrity, download_url
 from archai.common.utils import download_and_extract_tar, extract_tar
-from archai.common.logger import Logger
 
-logger = Logger(source=__name__)
+
+from archai.common.common import logger
